requirements.txt
psycopg2-binary>=2.9.9
pgvector>=0.3.0
python-dotenv>=1.0.1
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
anthropic>=0.39.0
fastmcp>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
